requirements.txt
aiohttp==3.6.2
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
discord.py==1.3.1
dj-database-url==0.5.0
Django==2.1.5
giphy-client==1.0.0
gunicorn==19.9.0
idna==2.8
multidict==4.7.4
psycopg2-binary==2.7.7
python-dateutil==2.8.1
python-dotenv==0.10.5
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
websockets==8.1
whitenoise==4.1.2
wincertstore==0.2
yarl==1.4.2
